涉及到时间的参数全部调整为 int64

+++ json.go
@@ -12,7 +12,7 @@ func (this *Session) MarshalJSON(key string, obj interface{}) *Result {
 	return this.SET(key, string(value))
 }
 
-func (this *Session) MarshalJSONEx(key string, seconds int, obj interface{}) *Result {
+func (this *Session) MarshalJSONEx(key string, seconds int64, obj interface{}) *Result {
 	value, err := json.Marshal(obj)
 	if err != nil {
 		return result(nil, err)
